config dataclass
